refactor(tui): move recurrence-rule helpers to event_form_rrule.go
The four RRULE helpers (parseRecurrenceRule, parseRRuleUntil,
formatRRuleUntil, rruleParam) are not construction-only code. The save
path (buildRecurrenceRule in event_form_save.go) calls
formatRRuleUntil, and the edit-instance prefill calls rruleParam and
parseRecurrenceRule. Their placement under event_form_new.go misdescribed
their role as new-form code. Co-locate them in a dedicated file so the
file names describe what each file holds.

Pure code motion: bodies, signatures, and comments are unchanged. All
imports in event_form_new.go stay in use, so only the new file gains an
import block (strings, time).
